MAJOR FIXES
I found that the program performed better when the number of survivors are increased, but I discovered a major problem with calculating the index to put the offspring into. Turns out, sometimes, the program has one of the parents as the same array as the offspring, which lead to a very confusing bug.

I also wanted to make it so that the whole path is stored because I think that it is better if I do it that way, so I changed a bunch of things in order to make that happen.

I renamed the classes to make more sense.

Also, I changed some little things that the IDE suggested.